Create HarvardX-Data-Science.Rproj
I'm figuring out how to implement version control, which I'm doing for the first time ever!

So far, I've created a repository in which to store R scripts from the HarvardX DS program. I created an R project on RStudio and I think I've successfully stored the file in said repository. Then I made this branch (branch-1) and moved the change to it instead of making the commit on the main branch. What's next I guess is to commit the changes to this branch.

There.